requirements.txt
about-time==4.2.1
absl-py==1.3.0
alive-progress==3.1.4
astunparse==1.6.3
autograd==1.6.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cma==2.7.0
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
filelock==3.12.2
flatbuffers==22.9.24
fonttools==4.37.4
future==0.18.3
gast==0.4.0
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
imageio==2.22.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.6.1
networkx==2.8.7
nose2==0.12.0
nose2-html-report==0.6.0
numpy==1.23.4
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.3
Pillow==9.2.0
platformdirs==3.10.0
plotly==5.15.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.3.0
pymoo==0.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.11.1
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tifffile==2022.10.10
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.24.2
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
